nextjs-platform/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
aiosqlite==0.20.0
asyncpg==0.30.0
pydantic==2.9.2
python-dotenv==1.0.1
greenlet==3.1.1
